+++ Assets/Scripts/HealthTextView.cs
@@ -7,8 +7,8 @@ public class HealthTextView : HealthView
     [SerializeField] private char _separatorSign = '/';
     [SerializeField] private TextMeshProUGUI _healthText;
 
-    protected override void UpdateView(int health, int maxHealth)
+    protected override void UpdateView(float health, float maxHealth)
     {
-        _healthText.text = _healthTextValue + health + _separatorSign + maxHealth;
+        _healthText.text = _healthTextValue + (int)health + _separatorSign + (int)maxHealth;
     }
 }
